+++ R/elt-methods.R
@@ -37,10 +37,10 @@ setMethod("elt", "EL", function(object,
       calibrate, alpha, out$statistic, length(par), par, object, control
     )
     return(new("ELT",
-      alpha = alpha, cv = unname(cal["cv"]), rhs = par, lhs = h$l,
-      calibrate = calibrate, optim = optim, logp = setNames(out$logp, onames),
-      logl = out$logl, loglr = out$loglr, statistic = out$statistic,
-      pval = unname(cal["pval"])
+      optim = optim, logp = setNames(out$logp, onames), logl = out$logl,
+      loglr = out$loglr, statistic = out$statistic, pval = unname(cal["pval"]),
+      cv = unname(cal["cv"]), rhs = par, lhs = h$l, alpha = alpha,
+      calibrate = calibrate
     ))
   }
   # Proceed with chi-square calibration for non-NULL `lhs`
@@ -57,10 +57,11 @@ setMethod("elt", "EL", function(object,
   optim <- validate_optim(out$optim)
   names(optim$par) <- pnames
   new("ELT",
-    alpha = alpha, cv = qchisq(1 - alpha, df = nrow(h$l)), rhs = h$r, lhs = h$l,
-    calibrate = calibrate, optim = optim, logp = setNames(out$logp, onames),
-    logl = out$logl, loglr = out$loglr, statistic = out$statistic,
-    pval = pchisq(out$statistic, df = nrow(h$l), lower.tail = FALSE)
+    optim = optim, logp = setNames(out$logp, onames), logl = out$logl,
+    loglr = out$loglr, statistic = out$statistic,
+    pval = pchisq(out$statistic, df = nrow(h$l), lower.tail = FALSE),
+    cv = qchisq(1 - alpha, df = nrow(h$l)), rhs = h$r, lhs = h$l, alpha = alpha,
+    calibrate = calibrate
   )
 })
 
@@ -108,10 +109,10 @@ setMethod("elt", "QGLM", function(object,
       calibrate, alpha, out$statistic, length(par), par, object, control
     )
     return(new("ELT",
-      alpha = alpha, cv = unname(cal["cv"]), rhs = h$r, lhs = h$l,
-      calibrate = calibrate, optim = optim, logp = setNames(out$logp, onames),
-      logl = out$logl, loglr = out$loglr, statistic = out$statistic,
-      pval = unname(cal["pval"])
+      optim = optim, logp = setNames(out$logp, onames), logl = out$logl,
+      loglr = out$loglr, statistic = out$statistic, pval = unname(cal["pval"]),
+      cv = unname(cal["cv"]), rhs = h$r, lhs = h$l, alpha = alpha,
+      calibrate = calibrate
     ))
   }
   stopifnot(
@@ -127,10 +128,11 @@ setMethod("elt", "QGLM", function(object,
   optim <- validate_optim(out$optim)
   names(optim$par) <- pnames
   new("ELT",
-    alpha = alpha, cv = qchisq(1 - alpha, df = nrow(l)), rhs = h$r, lhs = h$l,
-    calibrate = calibrate, optim = optim, logp = setNames(out$logp, onames),
-    logl = out$logl, loglr = out$loglr, statistic = out$statistic,
-    pval = pchisq(out$statistic, df = nrow(l), lower.tail = FALSE)
+    optim = optim, logp = setNames(out$logp, onames), logl = out$logl,
+    loglr = out$loglr, statistic = out$statistic,
+    pval = pchisq(out$statistic, df = nrow(l), lower.tail = FALSE),
+    cv = qchisq(1 - alpha, df = nrow(l)), rhs = h$r, lhs = h$l, alpha = alpha,
+    calibrate = calibrate
   )
 })
 
@@ -171,10 +173,10 @@ setMethod("elt", "SD", function(object,
     names(optim$par) <- pnames
     cal <- calibrate(calibrate, alpha, out$statistic, 1L, par, object, control)
     return(new("ELT",
-      alpha = alpha, cv = unname(cal["cv"]), rhs = h$r, lhs = h$l,
-      calibrate = calibrate, optim = optim, logp = setNames(out$logp, onames),
-      logl = out$logl, loglr = out$loglr, statistic = out$statistic,
-      pval = unname(cal["pval"])
+      optim = optim, logp = setNames(out$logp, onames), logl = out$logl,
+      loglr = out$loglr, statistic = out$statistic, pval = unname(cal["pval"]),
+      cv = unname(cal["cv"]), rhs = h$r, lhs = h$l, alpha = alpha,
+      calibrate = calibrate
     ))
   }
   stopifnot(
@@ -194,10 +196,11 @@ setMethod("elt", "SD", function(object,
   optim <- validate_optim(out$optim)
   names(optim$par) <- pnames
   new("ELT",
-    alpha = alpha, cv = qchisq(1 - alpha, df = 1L), rhs = h$r, lhs = h$l,
-    calibrate = calibrate, optim = optim, logp = setNames(out$logp, onames),
-    logl = out$logl, loglr = out$loglr, statistic = out$statistic,
-    pval = pchisq(out$statistic, df = 1L, lower.tail = FALSE)
+    optim = optim, logp = setNames(out$logp, onames), logl = out$logl,
+    loglr = out$loglr, statistic = out$statistic,
+    pval = pchisq(out$statistic, df = 1L, lower.tail = FALSE),
+    cv = qchisq(1 - alpha, df = 1L), rhs = h$r, lhs = h$l, alpha = alpha,
+    calibrate = calibrate
   )
 })